backend/requirements.txt
fastapi
uvicorn
google-generativeai
duckduckgo-search
beautifulsoup4
requests
weasyprint
pydantic
python-dotenv
markdown
mcp
